+++ package/src/firestoreAdmin/dbAdmin.js
@@ -43,7 +43,9 @@ const dbAdmin = (_ => {
 *   object|null (the value of the object, or that it's not there) when the predicate holds
 *   false if timed out
 */
-function eventually(docPath, p, timeoutMs) {    // (string, (object|null) => boolean, ms?) => Promise of object|null|false
+function eventually(docPath, p, timeoutMs) {    // (string, (object|null) => boolean, ms?) => Promise of object|null|undefined
+  // Undocumented: if 'p' omitted, just wait for any document
+  p = p || (o => !!o);
 
   return new Promise( (resolve) => {
     let unsub;
@@ -53,7 +55,7 @@ function eventually(docPath, p, timeoutMs) {    // (string, (object|null) => boo
     //
     if (timeoutMs) {
       timer = setTimeout( _ => {
-        resolve(false);   // timed out
+        resolve(undefined);   // timed out    // tbd. best API style for conveying timeout, still in flux
         unsub();
       }, timeoutMs);
     }
@@ -63,14 +65,40 @@ function eventually(docPath, p, timeoutMs) {    // (string, (object|null) => boo
 
       if (p(o)) {
         if (timer) clearTimeout(timer);
-        unsub();
         resolve(o);
+        unsub();
       }
     });
   })
 }
 
+/*
+* Listener for a collection.
+*
+* Note: Setting up the first listener to a collection carries a 300..320 ms penalty. By pre-warming the listening
+*     (if the test application sets up the listener in, say, 'beforeAll'), we can make the test times *seem* faster,
+*     dropping from ~570..620ms to ~280ms. However, this is a VISUAL TRICK since the overall execution time remains
+*     the same - now the delay simply is moved to the test preparation, instead of test execution.
+*
+*     Ideally, setting up a Firestore listener (with emulator) could be sped up by Firebase. It likely doesn't need to
+*     take ~300ms? #contribute: ask them
+*/
+function listener_EXP(collPath) {    // (string) => (string, ((object|null) => boolean)?)) => Promise of object|null
+
+  // Dummy listener that may be executed already in 'beforeAll' (depends on the test application, of course).
+  //
+  const unsub = dbAdmin.doc(`${collPath}/...`).onSnapshot( ss => {} );
+  unsub();
+
+  return (docId, p, timeoutMs) => {
+    return eventually(`${collPath}/${docId}`, p, timeoutMs);
+  }
+}
+
 export {
   dbAdmin,
-  eventually
+  eventually,
+
+  // EXPERIMENTAL
+  listener_EXP
 }